вынесет из main информацию с карточками в characters
