+++ src/pages/Characters/Characters.module.scss
@@ -1,1 +1,18 @@
-.root {}
+.cardSection {
+  display: flex;
+  justify-content: center;
+  padding: 24px 0;
+  background-color: #F7F7F7;
+}
+
+.cardTitle {
+  text-align: center;
+  margin-bottom: 32px;
+}
+
+.cardWrapper {
+  display: grid;
+  grid-template-columns: 1fr 1fr 1fr;
+  margin-bottom: 16px;
+}
+
